Changed  x axis

+++ app/src/main/java/com/example/first_and_goal/PointValue.java
@@ -1,17 +1,19 @@
 package com.example.first_and_goal;
 
+import java.text.SimpleDateFormat;
+
 public class PointValue {
-    long xValue;
+     int xValue;
     int yValue;
 
     public PointValue(){}
 
-    public PointValue(long xValue, int yValue){
+    public PointValue(int xValue, int yValue){
         this.xValue = xValue;
         this.yValue = yValue;
     }
 
-    public long getxValue(){
+    public int getxValue(){
         return xValue;
     }
